requirements.txt
aiogram==2.25.1
Django==5.0.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
